Update equal btn function for futher calculations
I set answer to num1 variable when the button is
clicked because answer value will be taken as the
num1 value if user wants to do further
calculations.

I also set num2 to 0 because when the user clicks
a number button after pressing the equal we want
to assign that number to num2 variable.