Fix the bug of number getting concatenated
After getting an answer it is updated into the
display. Then when user enters a new number it got
concatenated to the display number which is the
answer and also num1.

This is not the normal behaivour of calculators.
In this case it should display a new number and
assign that new number to num1 when user enters a
number, not concatenated to the current answer.