Fix the error of backspace on the answer
An error was shown when the backspace was clicked
after getting an answer. This is due to num1 being
a number and it couldn't use the substring method
without converting it to a string.

And it also didn't erase digits after an operator
was selected. Changed the if...else statement to
address that.