Update 0 button function
Earlier when the user hits the 0 button when there
was no value for num1 and num2, display shrak
because there was no value to show.

Here I added string 0 so when the user hits 0
again it won't concatenate to the current value
because variable is true because of the string 0.